Initial refactor for related pub create
